+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Version 7.x *(unreleased)*
 
+## Version 7.0.3 *(unreleased)*
+* Add support for ARM64 Windows [#315](https://github.com/node-gradle/gradle-node-plugin/issues/315)
+
 ## Version 7.0.2 *(2024-02-02)*
 * Prevent misconfigured `workDir` from removing all unrelated files [#297](https://github.com/node-gradle/gradle-node-plugin/issues/297)